requirements.txt
requests>=2.31.0
lxml>=5.0.0
python-dotenv>=1.0.0
psycopg-binary>-2.9.9
pathlib2>=2.3.7
